Draft scripts for running and debugging java cucumebr tests locally
These are instructions for running java cucumber tests locally or in a debugger.

Run scripts/localSetup.sh to setup the environment.
The script will require as input the full path to java-algorand-sdk.
When prompted by the script, enter the full path to java-algorand-sdk in the command line and press Enter.
The provided location will be used as the source code under test of the java sdk.
Note: run localSetup.sh without any flags. The behavior is undefined with flags.

Run debugTest.sh to run the tests locally with the following flags.
./scripts/debugTest.sh --java --cross

To debug in a debugger, comment the line:
Then run the script. When prompted for: Press enter when debugging finished ...
Run the tests in the debugger (RunCucumberTest.java).
Press Enter to shutdown the testing environment.

Notes:
In java_cucumber/src/test/java/java_cucumber/Stepdefs.java
If KMD_DIR environment variable is not set properly, the follwing change might be needed:
Change from:
data_dir_path += System.getenv("KMD_DIR") + "/";
To:
data_dir_path += "kmd-v0.5" + "/";